avcodec/hevc_ps: further constrain allowed num_ref_loc_offsets values
The spec says: "The value of num_ref_loc_offsets shall be in the range of 0 to
vps_max_layers_minus1, inclusive".

+++ libavcodec/hevc_ps.h
@@ -314,7 +314,7 @@ typedef struct HEVCPPS {
     uint8_t poc_reset_info_present_flag;
     uint8_t pps_infer_scaling_list_flag;
     uint8_t pps_scaling_list_ref_layer_id;
-    uint16_t num_ref_loc_offsets;
+    uint8_t num_ref_loc_offsets;
     uint8_t ref_loc_offset_layer_id[64];
     uint8_t scaled_ref_layer_offset_present_flag[64];
     int8_t scaled_ref_layer_left_offset[64];
